requirements.txt
requests>=2.32.3
PyYAML>=6.0.2
python-dateutil>=2.9.0.post0
google-auth>=2.35.0
